Issue #296: Cleanup and Implemetation
