+++ src/backend/hanningFunction.cpp
@@ -1,6 +1,4 @@
 #define _USE_MATH_DEFINES
-#include <cmath>
-#include <stdexcept>
 #include "hanningFunction.h"
 
 std::vector<double> hanningFunction(int windowSize) {
